# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Small-Server",
    "version": "1.1.0",
    "description": "A tiny REST API testing server — full CRUD on any JSON payload, SQLite-backed. Built with Go."
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Item": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "payload": { "description": "Whatever JSON you sent, stored verbatim" },
          "scope": { "type": "string", "enum": ["public", "secure"] },
          "created_at": { "type": "string" }
        }
      },
      "Cleared": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "cleared" },
          "scope": { "type": "string" },
          "removed": { "type": "integer", "description": "count of items deleted" }
        }
      },
      "Deleted": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "deleted" },
          "id": { "type": "integer" }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "detail": { "type": "string" } }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["meta"],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string", "example": "ok" } }
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": ["meta"],
        "summary": "Lightweight in-memory telemetry (resets on restart)",
        "responses": {
          "200": {
            "description": "stats snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uptime_seconds": { "type": "integer" },
                    "started_at": { "type": "string", "format": "date-time" },
                    "total_requests": { "type": "integer" },
                    "by_method": { "type": "object", "additionalProperties": { "type": "integer" } },
                    "by_status": { "type": "object", "additionalProperties": { "type": "integer" } },
                    "last_request": { "type": "string", "format": "date-time" },
                    "latency": {
                      "type": "object",
                      "description": "Latency over the last 100 requests",
                      "properties": {
                        "avg_ms": { "type": "number" },
                        "min_ms": { "type": "number" },
                        "max_ms": { "type": "number" },
                        "p95_ms": { "type": "number" },
                        "samples": { "type": "integer" }
                      }
                    },
                    "items_total": { "type": "integer" },
                    "items_public": { "type": "integer" },
                    "items_secure": { "type": "integer" },
                    "db_size_bytes": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/items": {
      "get": {
        "tags": ["public"],
        "summary": "List all public items",
        "responses": {
          "200": {
            "description": "array of items",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Item" } } } }
          }
        }
      },
      "post": {
        "tags": ["public"],
        "summary": "Create — store ANY JSON payload (no auth)",
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "description": "Any valid JSON — object, array, string, number, boolean, or null" } } } },
        "responses": {
          "201": { "description": "created item", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } }
        }
      },
      "delete": {
        "tags": ["public"],
        "summary": "Clear all — delete every public item",
        "responses": {
          "200": { "description": "cleared", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cleared" } } } }
        }
      }
    },
    "/public/items/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
      "get": {
        "tags": ["public"],
        "summary": "Read one public item by id",
        "responses": {
          "200": { "description": "item", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } },
          "404": { "description": "not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "tags": ["public"],
        "summary": "Update — replace the payload of a public item",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "description": "Any valid JSON — replaces the existing payload" } } } },
        "responses": {
          "200": { "description": "updated item", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } },
          "404": { "description": "not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["public"],
        "summary": "Delete one public item by id",
        "responses": {
          "200": { "description": "deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Deleted" } } } },
          "404": { "description": "not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/secure/items": {
      "get": {
        "tags": ["secure"],
        "summary": "List all secure items",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": { "description": "array of items", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Item" } } } } },
          "401": { "description": "missing/invalid API key" }
        }
      },
      "post": {
        "tags": ["secure"],
        "summary": "Create — store ANY JSON payload",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "description": "Any valid JSON" } } } },
        "responses": {
          "201": { "description": "created item", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } },
          "401": { "description": "missing/invalid API key" }
        }
      },
      "delete": {
        "tags": ["secure"],
        "summary": "Clear all — delete every secure item",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": { "description": "cleared", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cleared" } } } },
          "401": { "description": "missing/invalid API key" }
        }
      }
    },
    "/secure/items/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
      "get": {
        "tags": ["secure"],
        "summary": "Read one secure item by id",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": { "description": "item", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } },
          "404": { "description": "not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "missing/invalid API key" }
        }
      },
      "put": {
        "tags": ["secure"],
        "summary": "Update — replace the payload of a secure item",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "description": "Any valid JSON — replaces the existing payload" } } } },
        "responses": {
          "200": { "description": "updated item", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } },
          "404": { "description": "not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "missing/invalid API key" }
        }
      },
      "delete": {
        "tags": ["secure"],
        "summary": "Delete one secure item by id",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": { "description": "deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Deleted" } } } },
          "404": { "description": "not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "missing/invalid API key" }
        }
      }
    }
  }
}